backend/app/subtitercmd/clippercmd/settings.py
# clipper/clippercmd/settings.py

from dotenv import load_dotenv
import os


class Settings:
    """Loads and holds application settings."""
    def __init__(self):
        # Load environment variables from a .env file
        load_dotenv()
        self.OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
        self.DEEPGRAM_API_KEY = os.getenv("DEEPGRAM_API_KEY")
        self.AWS_REGION = os.getenv("AWS_REGION")
        self.AWS_ACCESS_KEY_ID = os.getenv("AWS_ACCESS_KEY_ID")
        self.AWS_SECRET_ACCESS_KEY = os.getenv("AWS_SECRET_ACCESS_KEY")